+++ tests/html/CompoundTest.php
@@ -7,9 +7,9 @@ require __DIR__ . '../../../vendor/autoload.php';
 use DBlackborough\Quill\Render as QuillRender;
 
 /**
- * Composite tests, multiple attributes and complex deltas
+ * Compound tests, multiple attributes and more complex deltas
  */
-final class CompositeTest extends \PHPUnit\Framework\TestCase
+final class CompoundTest extends \PHPUnit\Framework\TestCase
 {
     private $delta_multiple_attributes = '{
         "ops":[
@@ -53,7 +53,37 @@ final class CompositeTest extends \PHPUnit\Framework\TestCase
         ]
     }';
 
+    private $delta_multiple_unknown_attributes_image = '{
+        "ops": [
+            {
+                "insert": "Text 1 "
+            }, 
+            {
+                "attributes": {
+                    "bold": true
+                },
+                "insert": "assumenda"
+            }, 
+            {
+                "insert": " Text 2.\n\n"
+            }, 
+            {
+                "attributes": {
+                    "width": "214",
+                    "style": "display: inline; float: right; margin: 0px 0px 1em 1em;"
+                },
+                "insert": {
+                    "image": "data:image/png;base64,ImageDataOmittedforSize"
+                }
+            }, 
+            {
+                "insert": "\n\nText 3."
+            }
+        ]
+    }';
+
     private $expected_multiple_attributes = "<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed efficitur nibh tempor augue lobortis, nec eleifend velit venenatis. Nullam fringilla dui eget lectus mattis tincidunt. Donec sollicitudin, lacus sed luctus ultricies, <s><em>quam sapien </em></s><s>sollicitudin</s> quam, nec auctor eros felis elementum quam. Fusce vel mollis enim. <strong>Sed ac augue tincidunt,</strong> cursus urna a, tempus ipsum. Donec pretium fermentum erat a <u>elementum</u>. In est odio, mattis sed dignissim sed, porta ac nisl. Nunc et tellus imperdiet turpis placerat tristique nec quis justo. Aenean nisi libero, auctor a laoreet sed, fermentum vel massa. Etiam ultricies leo eget purus tempor dapibus. Integer ac sapien eros. Suspendisse convallis ex.</p>";
+    private $expected_multiple_unknown_attributes_image = '<p>Text 1 <strong>assumenda</strong> Text 2.</p><p><img src="data:image/png;base64,ImageDataOmittedforSize" width="214" style="display: inline; float: right; margin: 0px 0px 1em 1em;" /></p><p>Text 3.</p>';
 
     /**
      * Test a delta with multiple attributes
@@ -73,5 +103,29 @@ final class CompositeTest extends \PHPUnit\Framework\TestCase
         }
 
         $this->assertEquals($this->expected_multiple_attributes, $result, __METHOD__ . ' Multiple attributes failure');
+    }
+
+    /**
+     * Test a delta with multiple unknown attributes ona an image, attributes should be included as is
+     *
+     * @return void
+     * @throws \Exception
+     */
+    public function testUndefinedAttributesOnAnImage()
+    {
+        $result = null;
+
+        try {
+            $quill = new QuillRender($this->delta_multiple_unknown_attributes_image);
+            $result = $quill->render();
+        } catch (\Exception $e) {
+            $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
+        }
+
+        $this->assertEquals(
+            $this->expected_multiple_unknown_attributes_image,
+            $result,
+            __METHOD__ . ' - Undefined attributes on image failure'
+        );
     }
 }